added tests to services and VideoController

+++ HibernateTest/src/main/java/com/tv/dao/ChannelDao.java
@@ -7,11 +7,11 @@ import com.tv.model.Channel;
 public interface ChannelDao {
 	public List<Channel> findAll();
 
-	public Channel findById(Long id);
+	public Channel findById(Integer id);
 
 	public Channel save(Channel channel);
 
-	public void delete(Long id);
+	public void delete(Integer id);
 
 	public Channel findAllVideosByChannel(Integer id);
