+++ HibernateTest/src/main/java/com/tv/dao/VideoDao.java
@@ -8,9 +8,9 @@ public interface VideoDao {
 	
 	public List<Video> findAll();
 
-	public Video findById(Long id);
+	public Video findById(Integer id);
 
 	public Video save(Video video);
 
-	public void delete(Long id);
+	public void delete(Integer id);
 }
