+++ HibernateTest/src/main/java/com/tv/dao/services/VideoService.java
@@ -7,10 +7,10 @@ import com.tv.model.Video;
 public interface VideoService {
 	public List<Video> findAll();
 
-	public Video findById(Long id);
+	public Video findById(Integer id);
 
 	public Video save(Video video);
 
-	public void delete(Long id);
+	public void delete(Integer id);
 
 }
